Floyd-Warshall algorithm
Floyd-Warshall algorithm is used to find all pair shortest path problem from a given weighted graph. 

As a result of this algorithm, it will generate a matrix, which will represent the minimum distance from any node to all other nodes in the graph.

The time complexity of this algorithm is O(V3), here V is the number of vertices in the graph.